hex_tarball (hex_core v0.7.1)

Functions for creating and unpacking Hex tarballs.

Types

checksum()

contents()

filename()

files()

metadata()

tarball()

Functions

create ( Metadata, Files )

create ( Metadata, Files, Config )

Creates a package tarball.

Returns the binary of the tarball the "inner checksum" and "outer checksum". The inner checksum is deprecated in favor of the outer checksum.

Examples:

  > Metadata = #{<<"name">> => <<"foo">>, <<"version">> => <<"1.0.0">>},
  > Files = [{"src/foo.erl", <<"-module(foo).">>}],
  > hex_tarball:create(Metadata, Files).
  {ok, #{tarball => <<86,69,...>>,
         outer_checksum => <<40,32,...>>,
         inner_checksum => <<178,12,...>>}}

create_docs ( Files )

create_docs ( Files, _ )

Creates a docs tarball.

Examples:

  > Files = [{"doc/index.html", <<"Docs">>}],
  > hex_tarball:create_docs(Files).
  {ok, <<86,69,...>>}

format_byte_size ( Size )

format_checksum ( Checksum )

Returns base16-encoded representation of checksum.

format_error ( _ )

Converts an error reason term to a human-readable error message string.

unpack ( Tarball, Output )

unpack ( Tarball, Output, Config )

Unpacks a package tarball.
Remember to verify the outer tarball checksum against the registry checksum returned from hex_repo:get_package(Config, Package)`. Examples: ``` > hex_tarball:unpack(Tarball, memory). {ok,#{outer_checksum => <<...>>, contents => [{"src/foo.erl",<<"-module(foo).">>}], metadata => #{<<"name">> => <<"foo">>, ...}}} > hex_tarball:unpack(Tarball, "path/to/unpack"). {ok,#{outer_checksum => <<...>>, metadata => #{<<"name">> => <<"foo">>, ...}}}''

unpack_docs ( Tarball, Output )

unpack_docs ( Tarball, Output, Config )

Unpacks a documentation tarball.

Examples:

  > hex_tarball:unpack_docs(Tarball, memory).
  {ok, [{"index.html", <<"<!doctype>">>}, ...]}
 
  > hex_tarball:unpack_docs(Tarball, "path/to/unpack").
  ok