hex_core v0.6.1 hex_tarball

Functions for working with Hex tarballs.

Functions

create/2

create(metadata(), files()) :: {:ok, {tarball(), checksum()}}

Creates a package tarball.

Returns the binary of the tarball the "inner checksum" and "outer checksum". The inner checksum is deprecated in favor of the inner checksum.

Examples:

  > Metadata = #{<<"name">> => <<"foo">>, <<"version">> => <<"1.0.0">>},
  > Files = [{"src/foo.erl", <<"-module(foo).">>}],
  > hex_tarball:create(Metadata, Files).
  {ok, #{tarball => <<86,69,...>>,
         outer_checksum => <<40,32,...>>,
         inner_checksum => <<178,12,...>>}}

create_docs/1

create_docs(files()) :: {:ok, {tarball(), checksum()}}

Creates a docs tarball.

Examples:

  > Files = [{"doc/index.html", <<"Docs">>}],
  > hex_tarball:create_docs(Files).
  {ok, #{tarball => <<86,69,...>>,
         checksum => <<40,32,...>>}

format_checksum/1

format_checksum(checksum()) :: binary()

Returns base16-encoded representation of checksum.

format_error/1

format_error(term()) :: string()

Converts an error reason term to a human-readable error message string.

unpack/2

unpack(tarball(), :memory) ::
  {:ok,
   %{
     optional(:checksum) => checksum(),
     optional(:metadata) => metadata(),
     optional(:contents) => contents()
   }}
  | {:error, term()}

unpack(tarball(), filename()) ::
  {:ok, %{optional(:checksum) => checksum(), optional(:metadata) => metadata()}}
  | {:error, term()}

Unpacks a package tarball.

Remember to verify the outer tarball checksum against the registry checksum returned from hex_repo:get_package(Config, Package).

Examples:

  > hex_tarball:unpack(Tarball, memory).
  {ok,#{outer_checksum => <<...>>,
        contents => [{"src/foo.erl",<<"-module(foo).">>}],
        metadata => #{<<"name">> => <<"foo">>, ...}}}
 
  > hex_tarball:unpack(Tarball, "path/to/unpack").
  {ok,#{outer_checksum => <<...>>,
        metadata => #{<<"name">> => <<"foo">>, ...}}}`''